Move jira tasks to sentry.integrations.jira.tasks
