+++ src/sentry/tasks/integrations/__init__.py
@@ -26,7 +26,6 @@ __all__ = (
     "migrate_repo",
     "should_comment_sync",
     "sync_assignee_outbound",
-    "sync_metadata",
     "sync_status_inbound",
     "sync_status_outbound",
     "update_comment",
@@ -54,7 +53,6 @@ from .create_comment import create_comment
 from .kick_off_status_syncs_impl import kick_off_status_syncs
 from .migrate_repo import migrate_repo
 from .sync_assignee_outbound_impl import sync_assignee_outbound
-from .sync_metadata import sync_metadata
 from .sync_status_inbound import sync_status_inbound
 from .sync_status_outbound import sync_status_outbound
 from .update_comment import update_comment
