+++ tests/sentry/integrations/jira/test_sync_metadata.py
@@ -1,6 +1,6 @@
 import responses
 
-from sentry.tasks.integrations.sync_metadata import sync_metadata
+from sentry.integrations.jira.tasks import sync_metadata
 from sentry.testutils.cases import TestCase
 from sentry.testutils.silo import control_silo_test
